+++ src/primitive/defs.rs
@@ -954,6 +954,10 @@ primitive!(
     /// ex: °⊂ [1 2 3 4]
     /// ex: °⊂ [1_2 3_4 5_6]
     ///
+    /// Numeric subscripted [join] joins that many arrays.
+    /// ex: ⊂₃ 1 2_3_4 5_6
+    /// ex: ⊂₄ "al" "ig" "at" "or"
+    ///
     /// [join]ing to the front of an array is a bit slower than [join]ing to the back because it requires all the existing rows to be shifted.
     ///
     /// [join]'s glyph is `⊂` because it kind of looks like a magnet pulling its two arguments together.