+++ changelog.md
@@ -5,6 +5,7 @@ Uiua is not yet stable.
 ## 0.17.0 - 2025-??-??
 This version is not yet released. If you are reading this on the website, then these changes are live here.
 ### Language
+- **Breaking Change** - Numeric subscripted [`join ⊂`](https://uiua.org/docs/join) now joins that many arrays
 - Add array pack syntactic sugar. This lets you write code like `[⊃(+|×)]` as `⊃[+|×]`.
 
 ## 0.16.0 - 2025-05-18